Some major cleaning up.
It all started with some minor refactorings based on IDEA's recommendations, but then it turned into a big fix for type safety. A lot of code had to be made specific for FCAElementImplementation -- not nice, but the only typesafe option at the moment.

git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1996 267469b6-68fb-4578-8a4b-5844dc4013a7

+++ toscanaj/source/net/sourceforge/toscanaj/view/diagram/DiagramEditingView.java
@@ -854,7 +854,7 @@ public class DiagramEditingView extends JPanel implements EventBrokerListener {
                 final String baseStr = diagramTitle.substring(0, index1).trim();
                 try {
                     final Integer num = new Integer(numStr);
-                    copiedDiagramTitle = baseStr + " (" + (num.intValue() + 1)
+                    copiedDiagramTitle = baseStr + " (" + (num + 1)
                             + ")";
                 } catch (final NumberFormatException e) {
                     // ignore this exception because if whatever is in
